import-sources: cast statvfs fields so the crate compiles on macOS
Third instance of the same defect class: statvfs field widths differ by
platform (u64 on Linux, u32 on Darwin), so multiplying blocks_available()
by fragment_size() unconverted compiles only on Linux.

Found by building the workspace on a mac after the pdf rlimit fix
unblocked it -- this crate was hidden behind that failure.

A sweep of every statvfs field use in the workspace now shows all four
sites casting explicitly (this one, solstone-core::check, the
fit_report::free_bytes precedent, and doctor's disk_space via f64), so
the class is closed rather than fixed one site at a time.

+++ core/crates/solstone-core-import-sources/src/archive.rs
@@ -393,7 +393,11 @@ fn extract_archive(
             detail: error.to_string(),
         }
     })?;
-    let available = stat.blocks_available().saturating_mul(stat.fragment_size());
+    // Cast explicitly: `statvfs`'s field widths differ by platform (u64 on Linux,
+    // u32 on Darwin), so multiplying them unconverted only compiles on Linux.
+    // Same defect class as `solstone-core::check` and the precedent in
+    // `solstone-core-local::install::fit_report::free_bytes`.
+    let available = (stat.blocks_available() as u64).saturating_mul(stat.fragment_size() as u64);
     if available < required {
         return cleanup_extraction(
             path,
